refactor: context api 분리
